third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/flex-shrink-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'flex-shrink' to a number: -3.14
  assert_approx_equals: expected 0 +/- 0.000001 but got -3.140000104904175
Harness: the test ran to completion.
